refactor(ui-react): remove unused imports
