+++ packages/ui-react/src/modules/Mode/index.js
@@ -1,2 +1,2 @@
 export { MODE } from './mode.constants'
-export { reducer as mode } from './mode.reducer'
+export { reducer as mode, setMode } from './mode.reducer'
